% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share-creation.R
\name{create_share}
\alias{create_share}
\title{Creates shares. Option for gold contribution or commodity contribution.}
\usage{
create_share(
  url.wallet = "http://127.0.0.1:63079/json_rpc",
  commodity.id,
  contrib.type,
  contrib.quantity,
  contrib.transaction.height,
  contrib.nonce,
  contrib.investor,
  ...
)
}
\arguments{
\item{url.wallet}{TODO}

\item{commodity.id}{TODO}

\item{contrib.type}{TODO}

\item{contrib.quantity}{TODO}

\item{contrib.transaction.height}{TODO}

\item{contrib.nonce}{TODO}

\item{contrib.investor}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
