% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share-creation.R
\name{create_share_item_groups}
\alias{create_share_item_groups}
\title{Creates two item groups. One for shares issued to investors who contribute commodity; one for gold contributors.}
\usage{
create_share_item_groups(
  url.wallet = "http://127.0.0.1:63079/json_rpc",
  commodity.id,
  ...
)
}
\arguments{
\item{url.wallet}{TODO}

\item{commodity.id}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
