% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_commodity_contribs}
\alias{get_commodity_contribs}
\title{Produces a data.frame containing instances of investors sending the bot the commodity type under its management}
\usage{
get_commodity_contribs(
  url.townforged = "http://127.0.0.1:18881/json_rpc",
  commodity.id,
  bot.account.id,
  ...
)
}
\arguments{
\item{url.townforged}{TODO}

\item{commodity.id}{TODO}

\item{bot.account.id}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
