% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_commodity_investors}
\alias{get_commodity_investors}
\title{Produces a data.frame containing instances of investors sending the bot the commodity under its management}
\usage{
get_commodity_investors(url.townforged, commodity.id, bot.account.id, ...)
}
\arguments{
\item{url.townforged}{TODO}

\item{commodity.id}{TODO}

\item{bot.account.id}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
