% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_custom_items}
\alias{get_custom_items}
\title{Produces a data.frame containing player-made items}
\usage{
get_custom_items(url.townforged = "http://127.0.0.1:18881/json_rpc", ...)
}
\arguments{
\item{url.townforged}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
