% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_gold_investors}
\alias{get_gold_investors}
\title{Produces a data.frame containing instances of investors sending the bot gold}
\usage{
get_gold_investors(url.townforged, bot.account.id, ...)
}
\arguments{
\item{url.townforged}{TODO}

\item{bot.account.id}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
