% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share-issuer-daemon.R
\name{share_issuer_daemon}
\alias{share_issuer_daemon}
\title{Bot for creating limit orders in response to market movements}
\usage{
share_issuer_daemon(
  url.townforged = "http://127.0.0.1:18881/json_rpc",
  url.wallet = "http://127.0.0.1:63079/json_rpc",
  commodity.id,
  bot.account.id,
  loop.sleep.time = 10,
  ...
)
}
\arguments{
\item{url.townforged}{TODO}

\item{url.wallet}{TODO}

\item{commodity.id}{TODO}

\item{bot.account.id}{TODO}

\item{loop.sleep.time}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
