% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-daemon.R
\name{mm_daemon}
\alias{mm_daemon}
\title{Bot for creating limit orders in response to market movements}
\usage{
mm_daemon(
  url.townforged = "http://127.0.0.1:18881/json_rpc",
  url.wallet = "http://127.0.0.1:63079/json_rpc",
  commodity.id,
  bot.account.id,
  order.quantity,
  price.increment,
  price.steps,
  init.midpoint.price,
  cancel.all.orders.and.exit = FALSE,
  loop.sleep.time = 10,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{url.townforged}{TODO}

\item{url.wallet}{TODO}

\item{commodity.id}{TODO}

\item{bot.account.id}{TODO}

\item{order.quantity}{TODO}

\item{price.increment}{TODO}

\item{price.steps}{TODO}

\item{init.midpoint.price}{TODO}

\item{loop.sleep.time}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
TODO
}
